node-side/requirements.txt
pylightning
uuid
requests